#142 Layout start refactoring
