+++ webapp/package.json
@@ -8,7 +8,7 @@
     "test": "jest --verbose",
     "prebuild": "eslint . --color",
     "build": "webpack --mode none -p --progress",
-    "prebuild-dev": "eslint . --color",
+    "prebuild-dev": "eslint . --color --fix",
     "build-dev": "webpack --mode none --devtool eval",
     "prebuild-dev-noTest": "eslint . --color --fix",
     "build-dev-noTest": "webpack --mode none --devtool source-map --env.noTest=true",
@@ -33,7 +33,7 @@
     "@types/material-ui": "^0.21.5",
     "@types/react": "^16.4.9",
     "@types/react-dom": "^16.0.7",
-    "awesome-typescript-loader": "^5.0.0-1",
+    "awesome-typescript-loader": "^5.2.1",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.6",
     "babel-plugin-transform-decorators-legacy": "^1.3.5",
@@ -61,7 +61,7 @@
     "source-map-loader": "^0.2.3",
     "style-loader": "^0.13.2",
     "typedoc": "^0.11.1",
-    "typescript": "^3.8.3",
+    "typescript": "^2.0.0",
     "url-join": "^4.0.0",
     "url-loader": "^0.5.8",
     "webpack": "^4.35.0",
