+++ webapp/redux/reducers/all.js
@@ -2,9 +2,9 @@ import { combineReducers } from 'redux';
 
 import general from './general';
 import notebook from './notebook';
-import flexlayout from './flexlayout';
+import layout from './layout';
 import errors from './errors';
 import drawer from './drawer';
 import topbar from './topbar';
 
-export default combineReducers({ general, notebook, flexlayout, errors, drawer, topbar });
+export default combineReducers({ general, notebook, layout, errors, drawer, topbar });
